see #1257_1

+++ src/main/java/com/pinganzhiyuan/service/ClientVersionService.java
@@ -4,8 +4,8 @@ import com.pinganzhiyuan.model.ClientVersion;
 
 public interface ClientVersionService {
 
-	ClientVersion getLatestVersionInfo(byte platformId, int channelId, String packageName);
+	ClientVersion getLatestVersionInfo(byte platformId, long channelId, String packageName);
 
-    ClientVersion getVersion(byte platformId, int channelId, String packageName, Integer versionCode);
+    ClientVersion getVersion(byte platformId, long channelId, String packageName, Integer versionCode);
 
 }
